pyproject.toml
[tool.poetry]
name = "aula"
version = "0.1.0"
description = ""
authors = ["Ricardo Borges <ricardoborges@ufg.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^1.1.3"
langchain-openai = "^1.1.1"
openai = "^2.11.0"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
